fix(pkg/kernelrelease): allow plus char as sublevel and extra separator
This is the case of distributions like Debian which applies a semantic
for which a separator can be the the plus character ('+').

For example, with this change, a release named '4.19+105+deb10u4~bpo9+1'
will be matched and the change will allow parsing by capturing the
resulting regular expression named capturing groups.

Signed-off-by: Massimiliano Giovagnoli <me@maxgio.it>